pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "termaite"
dynamic = ["version"]
description = "LLM-powered shell assistant with multi-agent architecture"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "term.ai.te Team" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Shells",
    "Topic :: Utilities",
]
dependencies = [
    "PyYAML>=6.0",
    "requests>=2.25.0",
    "colorama>=0.4.4",
]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/yourusername/termaite"
Repository = "https://github.com/yourusername/termaite"
Documentation = "https://github.com/yourusername/termaite#readme"
"Bug Tracker" = "https://github.com/yourusername/termaite/issues"

[project.scripts]
termaite = "termaite.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "isort>=5.0",
    "mypy>=1.0",
    "pre-commit>=2.0",
]

[tool.hatch.version]
path = "termaite/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["termaite"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
